PrepareSceneCopyForRanderingMasks
This is intended to use for preparing a scene copy for rendering masks. A full copy of the scene should be made. The script sets it up to render masks in red green and blue channels of an image with cycles. All scene's materials are changed to 'Masks.Black' material, then materials 'Masks.Red', 'Masks.Green' and 'Masks.Blue' should be assigned to desired objects.